requirements.txt
streamlit==1.29.0
pdfplumber==0.10.3
python-docx==1.1.0
openai==1.6.1
Flask==3.0.0
flask-cors==4.0.0
python-magic-bin==0.4.14
Pillow==10.1.0
pypdf==3.17.4
pdf2docx==0.5.6
python-dotenv==1.0.0
